add api user

+++ src/nguoi-dung/nguoi-dung.service.ts
@@ -399,18 +399,41 @@ export class NguoiDungService {
 
   // xóa người dùng
   async xoaNguoiDung(taiKhoan) {
-    let taiKhoanNum=Number(taiKhoan)
-    await this.prisma.nguoiDung.delete({
-      where:{tai_khoan:taiKhoanNum}
-    })
-    let data = {
-      "statusCode": 200,
-      "message": "Xóa Người Dùng thành công!",
-      "content": {
-        
+    if (taiKhoan) {
+      let taiKhoanNum = Number(taiKhoan)
+      let arrTaiKhoan = await this.prisma.nguoiDung.findMany({
+          select: { tai_khoan: true }
+      })
+      let index = arrTaiKhoan.findIndex((item) => {
+          return item.tai_khoan == taiKhoanNum
+      })
+      if (index == -1) {
+          let data = {
+              "statusCode": 400,
+              "message": "Yêu cầu không hợp lệ!",
+              "content": "Sai tài khoản!",
+          }
+          throw new HttpException(data, 400)
+      } else {
+          await this.prisma.nguoiDung.delete({
+              where: { tai_khoan: taiKhoanNum }
+          })
+          let data = {
+              "statusCode": 200,
+              "message": "Xóa người dùng thành công!",
+          }
+          return data
       }
-    }
-    return data
+
+  } else {
+      let data = {
+          "statusCode": 400,
+          "message": "Yêu cầu không hợp lệ!",
+          "content": "Thiếu tài khoản!",
+      }
+      throw new HttpException(data, 400)
+  }
+
     
     
   }
